small refactor: move builders, views to new common subpackage
