+++ internal/thelma/cli/commands/bee/common/views/bee.go
